refactor : 패키지 구조 형식에 맞게 리팩토링
