+++ src/main/java/com/sparta/deliverit/order/presentation/dto/response/OrderInfo.java
@@ -1,7 +1,7 @@
-package com.sparta.deliverit.order.dto.response;
+package com.sparta.deliverit.order.presentation.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.sparta.deliverit.order.entity.OrderStatus;
+import com.sparta.deliverit.order.domain.entity.OrderStatus;
 import lombok.Builder;
 import lombok.Getter;
