+++ src/test/java/com/sparta/deliverit/order/controller/OrderControllerV1Test.java
@@ -1,9 +1,10 @@
 package com.sparta.deliverit.order.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sparta.deliverit.order.dto.request.CreateOrderRequest;
-import com.sparta.deliverit.order.dto.request.OrderMenuRequest;
-import com.sparta.deliverit.order.entity.OrderStatus;
+import com.sparta.deliverit.order.presentation.controller.OrderControllerV1;
+import com.sparta.deliverit.order.presentation.dto.request.CreateOrderRequest;
+import com.sparta.deliverit.order.presentation.dto.request.OrderMenuRequest;
+import com.sparta.deliverit.order.domain.entity.OrderStatus;
 import jakarta.servlet.ServletException;
 import jakarta.validation.ConstraintViolationException;
 import org.assertj.core.api.Assertions;
